[libc] Add support for picolibc
This allows projects to set the LIBC variable to either 'newlib' or
'picolibc' to select between these two libraries.

RT-Thread expects picolibc to have been built with TLS support, and
that requires a small change to the project linker script to set up
TLS symbols, inserting this section:

    . = ALIGN(8);
    .tdata : {
	*(.tdata)
	*(.tdata.*)
	*(.gnu.linkonce.td.*)
    }

    .tbss (NOLOAD) : {
	*(.tbss .tbss.* .gnu.linkonce.tb.*)
	*(.tcommon)
    }

    __tdata_source = ADDR(.tdata);
    __tdata_size = SIZEOF(.tdata);
    __tbss_size = SIZEOF(.tbss);
    __tls_size = SIZEOF(.tdata) + SIZEOF(.tbss);

These should go with the section defining ROM contents as the '.tdata'
section will place initialization values for the TLS block in memory.

Signed-off-by: Keith Packard <keithp@keithp.com>